+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This release has an [MSRV][] of 1.82.
 ### Added
 
 * Support for the ACES2065-1 color space. ([#124][] by [@tomcur][])
+* A documentation example implementing `ColorSpace`. ([#130][] by [@tomcur][])
 
 ### Fixed
 
@@ -117,6 +118,7 @@ This is the initial release.
 [#124]: https://github.com/linebender/color/pull/124
 [#128]: https://github.com/linebender/color/pull/128
 [#129]: https://github.com/linebender/color/pull/129
+[#130]: https://github.com/linebender/color/pull/130
 
 [Unreleased]: https://github.com/linebender/color/compare/v0.2.2...HEAD
 [0.2.2]: https://github.com/linebender/color/releases/tag/v0.2.2