+++ CHANGELOG.md
@@ -8,10 +8,14 @@ Subheadings to categorize changes are `added, changed, deprecated, removed, fixe
 
 # Changelog
 
-The latest published Color release is [0.2.2](#022-2025-01-03) which was released on 2025-01-03.
-You can find its changes [documented below](#022-2025-01-03).
+The latest published Color release is [0.2.3](#023-2025-01-20) which was released on 2025-01-20.
+You can find its changes [documented below](#023-2025-01-20).
 
 ## [Unreleased]
+
+This release has an [MSRV][] of 1.82.
+
+## [0.2.3][] (2025-01-20)
 
 This release has an [MSRV][] of 1.82.
 
@@ -124,7 +128,8 @@ This is the initial release.
 [#135]: https://github.com/linebender/color/pull/135
 [#136]: https://github.com/linebender/color/pull/136
 
-[Unreleased]: https://github.com/linebender/color/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/linebender/color/compare/v0.2.3...HEAD
+[0.2.3]: https://github.com/linebender/color/releases/tag/v0.2.3
 [0.2.2]: https://github.com/linebender/color/releases/tag/v0.2.2
 [0.2.1]: https://github.com/linebender/color/releases/tag/v0.2.1
 [0.2.0]: https://github.com/linebender/color/releases/tag/v0.2.0